Add level counter to flash

+++ firmware/Eko_Badge/Levels/include/levels.h
@@ -10,7 +10,7 @@
 #define WINNER_BANNER "   L33t Hacker"
 
 extern void Levels_Init();
-extern uint8_t Get_Level();
+extern uint16_t Get_Level();
 extern void Set_Level(uint8_t newLevel);
 
 #endif
